+++ src/main/java/gift/service/WishListService.java
@@ -1,10 +1,11 @@
 package gift.service;
 
-import gift.dao.WishListDao;
 import gift.dto.WishListDto;
+import gift.entity.WishList;
+
 import gift.repository.ProductRepository;
+import gift.repository.WishListRepository;
 import gift.util.JwtUtil;
-import gift.domain.WishList;
 
 import java.util.stream.Collectors;
 
@@ -15,36 +16,35 @@ import java.util.List;
 @Service
 public class WishListService {
     
-    private WishListDao wishListDao;
+    private WishListRepository wishListRepository;
     private ProductRepository productRepository;
     private JwtUtil jwtUtil;
 
-    public WishListService(WishListDao wishListDao, ProductRepository productRepository, JwtUtil jwtUtil){
-        this.wishListDao =wishListDao;
+    public WishListService(WishListRepository wishListRepository, ProductRepository productRepository, JwtUtil jwtUtil) {
+        this.wishListRepository = wishListRepository;
         this.productRepository = productRepository;
         this.jwtUtil = jwtUtil;
     }
 
-    public List<WishListDto> findWishListById(String token){
+    public List<WishListDto> findWishListById(String token) {
 
         long userId = (long)jwtUtil.extractAllClaims(token).get("id");
-        List<WishList> wishlist = wishListDao.findWishListById(userId);
+        List<WishList> wishlist = wishListRepository.findProductIdsByMemberId(userId);
         return wishlist.stream()
         .map(WishListDto::fromEntity)
         .collect(Collectors.toList());
     }
 
-    public void addWishList(String token, long productId){
+    public void addWishList(String token, long productId) {
 
         long userId = (long)jwtUtil.extractAllClaims(token).get("id");
         productRepository.findById(productId);
-        wishListDao.insertWishList(productId, userId);
+        wishListRepository.save(new WishList(userId, productId));
 
     }
 
-    public void deleteWishList(String token, long productId){
-
+    public void deleteWishList(String token, long productId) {
         long userId = (long)jwtUtil.extractAllClaims(token).get("id");
-        wishListDao.deleteWishList(productId, userId);
+        wishListRepository.deleteById(wishListRepository.findId(userId, productId));;
     }
 }
